Move display and display_context to include folder
