+++ rviz_common/CMakeLists.txt
@@ -82,8 +82,8 @@ set(rviz_common_headers_to_moc
   include/rviz_common/frame_position_tracking_view_controller.hpp
   include/rviz_common/properties/quaternion_property.hpp
   include/rviz_common/visualizer_app.hpp
-  src/rviz_common/display.hpp
-  src/rviz_common/display_context.hpp
+  include/rviz_common/display.hpp
+  include/rviz_common/display_context.hpp
   # src/rviz_common/displays_panel.hpp
   src/rviz_common/failed_panel.hpp
   src/rviz_common/frame_manager.hpp
