+++ rviz_common/src/rviz_common/display.cpp
@@ -28,7 +28,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "./display.hpp"
+#include "rviz_common/display.hpp"
 
 #include <cstdio>
 #include <string>
@@ -55,7 +55,7 @@
 #include "rclcpp/time.hpp"
 #include "rviz_rendering/apply_visibility_bits.hpp"
 
-#include "./display_context.hpp"
+#include "rviz_common/display_context.hpp"
 #include "./panel_dock_widget.hpp"
 #include "./properties/property_tree_model.hpp"
 #include "./properties/status_list.hpp"
