+++ rviz_common/src/rviz_common/displays_panel.cpp
@@ -37,7 +37,7 @@
 #include <boost/bind.hpp>
 
 #include "./display_factory.hpp"
-#include "./display.hpp"
+#include "rviz_common/display.hpp"
 #include "./add_display_dialog.hpp"
 #include "rviz_common/properties/property.hpp"
 #include "./properties/property_tree_widget.hpp"
