+++ rviz_common/src/rviz_common/failed_display.hpp
@@ -31,7 +31,7 @@
 #ifndef RVIZ_COMMON__FAILED_DISPLAY_HPP_
 #define RVIZ_COMMON__FAILED_DISPLAY_HPP_
 
-#include "./display.hpp"
+#include "rviz_common/display.hpp"
 
 namespace rviz_common
 {
