+++ rviz_common/src/rviz_common/render_panel.cpp
@@ -51,7 +51,7 @@
 #include "rviz_rendering/render_window.hpp"
 
 // #include "./display.hpp"
-#include "./display_context.hpp"
+#include "rviz_common/display_context.hpp"
 #include "rviz_common/view_controller.hpp"
 #include "./viewport_mouse_event.hpp"
