+++ rviz_common/src/rviz_common/temp/default_plugins/displays/robot_model_display.cpp
@@ -52,7 +52,7 @@
 
 #include "../../robot/robot.hpp"
 #include "../../robot/tf_link_updater.hpp"
-#include "../../../display_context.hpp"
+#include "rviz_common/display_context.hpp"
 #include "../../../properties/float_property.hpp"
 #include "rviz_common/properties/property.hpp"
 #include "rviz_common/properties/string_property.hpp"
