+++ rviz_common/src/rviz_common/temp/default_plugins/displays/robot_model_display.hpp
@@ -35,7 +35,7 @@
 
 #include <OgreVector3.h>
 
-#include "../../../display.hpp"
+#include "rviz_common/display.hpp"
 
 namespace Ogre
 {
