+++ rviz_common/src/rviz_common/temp/default_plugins/displays/tf_display.hpp
@@ -41,7 +41,7 @@
 #include "tf2/time.h"
 
 #include "../../../selection/forwards.hpp"
-#include "../../../display.hpp"
+#include "rviz_common/display.hpp"
 
 namespace Ogre
 {
