+++ rviz_common/src/rviz_common/temp/default_plugins/tools/move_tool.cpp
@@ -29,7 +29,7 @@
 
 #include "move_tool.hpp"
 
-#include "../../../display_context.hpp"
+#include "rviz_common/display_context.hpp"
 #include "../../../render_panel.hpp"
 #include "../../../viewport_mouse_event.hpp"
 #include "../../../selection/selection_manager.hpp"
