+++ rviz_common/src/rviz_common/temp/default_plugins/view_controllers/orbit_view_controller.cpp
@@ -52,7 +52,7 @@
 # pragma GCC diagnostic pop
 #endif
 
-#include "../../../display_context.hpp"
+#include "rviz_common/display_context.hpp"
 #include "../../../load_resource.hpp"
 #include "../../../properties/bool_property.hpp"
 #include "../../../properties/float_property.hpp"
