+++ rviz_common/src/rviz_common/view_controller.cpp
@@ -52,7 +52,7 @@
 
 #include "rviz_rendering/render_window.hpp"
 
-#include "./display_context.hpp"
+#include "rviz_common/display_context.hpp"
 #include "./load_resource.hpp"
 #include "./properties/bool_property.hpp"
 // #include "./properties/enum_property.hpp"
